Restructure go code into packages
